// src/Api/Endpoints.ts
export enum Endpoints {
  SignIn = "login",
  SignUp = "user/register",
  ForgotPassword = "user/forgot-password",
  ChangePassword = "user/change-password",
  ResendOtp = "user/resend-otp",
  VerifyOtp = "user/verify-otp",

  //private router
  Me = "user/profile",
  UpdateProfile = "user/profile",
  FindPeople = "user/admire",
  FollowUser = "user/follow",
  UnFollowUser = "user/follow",
  Category = "category",
  FeedPosts = "user/post",
  Postview = "user/post/:id",
  LikePost = "user/post-like",
  ListComment = "user/post-comment",
  AddPost = "user/post",
  CreateLive = "user/broadcast",
  ListLive = "user/listlives",
  RecentPosts = "user/post/recent",
  UserList = "user/user",
  UserView = "user/user/:username",
  UserViewPost = "user/user",
  EndLive = "user/endlive",
  DeletePost = "user/post/:id",
  DeletePostTag = "user/post-tag",
  ReportPost = "user/report-post/:id",
  ReportCommentPost = "user/report-post/:id",
  JoinStream = "user/join-broadcast",
  StreamWatchCount = "user/get-view-count/:id",
  PublicPostView = "post/:id",
  PublicPostList = "post",
  ListAdmirning = "/user/admirning/:id",
  ListAdmirer = "/user/admirers/:id",
  // notification
  NotificationData = "user/notification",
  Notifications = "user/notification/:id",
  // hash tag
  PostHashtag = "user/post/:hashtag",
  // Pinpost
  PostPin = "user/post-pin",
  DeletePostPin = "user/post-pin/:id",
  // delete account
  DeleteAccount = "user/request-account-delete",
  //Firebase
  CreateLink = `https://firebasedynamiclinks.googleapis.com/v1/shortLinks?key=:key`,
  // Network
  AddNetwork = "user/network",
  NetworkListJoinAndNotJoin = `user/network`,
  NetworkFeedPosts = `user/network-post`,
  NetworkPostView = `user/network-post/:id`,
  PublicNetworkPostView = `network-post/:id`,
  NetworkPostLike = `user/post-like`,
  AddNetworkPostComment = `user/network-post`,
  DeleteNetworkPost = `user/network-post/:id`,
  NetworkMemberJoin = `/user/network-member-join/:invitation_code`,
  AddNetworkPost = `/user/network-post`,
  UpdateNetworkPost = `/user/network-post/:id`,
  UpdateNetwork = `/user/network/:id`,
  DeleteNetwork = `/user/network/:id`,
  NetworkGroupMemberList = `user/network-member`,
  UpdateNetworkGroupMember = `user/network-member/:id`,
  RemoveNetworkGroupMember = `user/network-member`,
  AcceptNetworkMember = `user/network-member-accept/:network_group_id`,
  RejectNetworkMember = `user/network-member-reject/:network_group_id`,
  PinNetworkPost = `user/network-post-pin`,
  UnpinNetworkPost = `user/network-post-pin/:network_post_id`,
  InviteNetworkMembers = `user/network-member-invite`,
  FollowNetwork = `user/network-follow`,
  UnFollowNetwork = `user/network-follow`,
  NewAgoraBearerToken = `user/agora/bearer-token`,
  UpdateLastActiveAt = `user/active`,
  UpdateAgoraChatUser = `user/agora/chat`,
  LiveInvite = `user/live-invite`,
  LiveAccept = `user/live-invite-accept/:id`,
  LiveReject = `user/live-invite-reject/:id`,
  GenerateRtmToken ="user/agora/rtm-token"
}